requirements.txt
discord.py
asyncpg
colorama
python-dotenv